Add WiFiManager support, sync settings to Blynk on bootup.
- Added WiFiManager so the wifi credentials don't need to
be in the code anymore. If the frame can't connect to the
saved wifi creds, it will open up an access point "Frames-AP"
for a phone to connect to and input wifi credentials. Upon
successful connection, it will save the settings.

- When the Frame connects to Blynk servers, it will sync all
the settings except power. That way, a Frame will always boot
off, but will have all the other settings in sync when each
color is turned on.